fixed bug in salary assignment, was not refetching new salary structure data
