+++ src/pages/Payroll/PayrollItems.js
@@ -7,6 +7,7 @@ import SalaryAssignment from "../../components/payroll-tabs/salary-assignment";
 import SalaryComponents from "../../components/payroll-tabs/salary-components";
 import SalaryStructure from "../../components/payroll-tabs/salary-structure";
 import { useAppContext } from "../../Context/AppContext";
+import axiosInstance from "../../services/api";
 
 const PayrollItems = () => {
   const [formType, setformType] = useState("");
@@ -15,6 +16,7 @@ const PayrollItems = () => {
   const [submitted, setsubmitted] = useState(false);
   const [path, setpath] = useState("/personal-details");
   const [editData, seteditData] = useState({});
+  const [data, setData] = useState([]);
 
   const { combineRequest } = useAppContext();
 
@@ -65,6 +67,21 @@ const PayrollItems = () => {
 
   useEffect(() => {
     fetchedCombineRequest();
+  }, []);
+
+  const fetchSalaryStructures = () => {
+    axiosInstance
+      .get("/api/salary-structure")
+      .then((res) => {
+        console.log(res);
+        setData(res.data.data);
+      })
+      .catch((error) => {
+        console.log(error?.response);
+      });
+  };
+  useEffect(() => {
+    fetchSalaryStructures();
   }, []);
   return (
     <>
@@ -118,8 +135,11 @@ const PayrollItems = () => {
           submitted={submitted}
           formValue={formValue}
         />
-        <SalaryStructure />
-        <SalaryAssignment />
+        <SalaryStructure
+          data={data}
+          fetchSalaryStructures={fetchSalaryStructures}
+        />
+        <SalaryAssignment salaryStructure={data} />
       </div>
       <FormModal
         editData={editData}
